chore: renamed component
